# apps/docs/public/openapi.yaml
openapi: '3.0.0'
info:
  version: v2.7.0/366f528
  title: Labrinth
  termsOfService: https://modrinth.com/legal/terms
  contact:
    name: Modrinth Support
    url: https://support.modrinth.com
    email: support@modrinth.com
  description: |
    This documentation doesn't provide a way to test our API. In order to facilitate testing, we recommend the following tools:

    - [cURL](https://curl.se/) (recommended, command-line)
    - [ReqBIN](https://reqbin.com/) (recommended, online)
    - [Postman](https://www.postman.com/downloads/)
    - [Insomnia](https://insomnia.rest/)
    - Your web browser, if you don't need to send headers or a request body

    Once you have a working client, you can test that it works by making a `GET` request to `https://staging-api.modrinth.com/`:

    ```json
    {
      "about": "Welcome traveler!",
      "documentation": "https://docs.modrinth.com",
      "name": "modrinth-labrinth",
      "version": "2.7.0"
    }
    ```

    If you got a response similar to the one above, you can use the Modrinth API!
    When you want to go live using the production API, use `api.modrinth.com` instead of `staging-api.modrinth.com`.

    ## Authentication
    This API has two options for authentication: personal access tokens and [OAuth2](https://en.wikipedia.org/wiki/OAuth).
    All tokens are tied to a Modrinth user and use the `Authorization` header of the request.

    Example:
    ```
...
    ```

...
    - those which create data (such as version creation)
...

...

...

...
servers:
...
components:
  parameters:
    ProjectIdentifier:
      name: id|slug
      in: path
      required: true
      description: The ID or slug of the project
      schema:
        type: string
        example: [AABBCCDD, my_project]
...
    VersionIdentifier:
      name: id
      in: path
      required: true
      description: The ID of the version
      schema:
        type: string
        example: [IIJJKKLL]
...
  schemas:
...
    EditableVersion:
      allOf:
        - $ref: '#/components/schemas/BaseVersion'
        - type: object
          properties:
            primary_file:
              type: array
              items:
                type: string
              example: [sha1, aaaabbbbccccddddeeeeffffgggghhhhiiiijjjj]
              description: The hash format and the hash of the new primary file
            file_types:
              type: array
              items:
                $ref: '#/components/schemas/EditableFileType'
              description: A list of file_types to edit
...
    EnvironmentEnum:
      type: string
      enum:
        - client_and_server
        - client_only
        - client_only_server_optional
        - singleplayer_only
        - server_only
        - server_only_client_optional
        - dedicated_server_only
        - client_or_server
        - client_or_server_prefers_both
        - unknown
      example: client_and_server
      description: 'The environment a project or version supports. For an explanation of each environment, see the blog post here: https://modrinth.com/news/article/new-environments/#new-system'
...
    EditableProject:
      allOf:
        - $ref: '#/components/schemas/ModifiableProject'
        - type: object
          properties:
            moderation_message:
              type: string
              description: The title of the moderators' message for the project
              nullable: true
            moderation_message_body:
              type: string
              description: The body of the moderators' message for the project
              nullable: true
    # Fields only available for project creation. https://github.com/modrinth/labrinth/blob/master/src/routes/project_creation.rs#L129-L197
...
    Project:
      type: object
      properties:
        id:
          type: string
          example: AABBCCDD
          description: The ID of the project, encoded as a base62 string
        team:
          type: string
          example: MMNNOOPP
          description: The ID of the team that has ownership of this project
        title:
          type: string
          description: The title or name of the project
        description:
          type: string
          description: A short sentence summarizing the project, no more than a sentence or two.
        body:
          type: string
          description: A long form description of the project
          example: A long body describing my project in detail
        status:
          type: string
          enum:
            [
              approved,
              archived,
              rejected,
              draft,
              unlisted,
              processing,
              withheld,
              scheduled,
              private,
              unknown,
            ]
          description: The status of the project
          example: approved
        project_type:
          type: string
          enum: [mod, modpack, resourcepack, shader]
          description: The project type of the project
          example: mod
        categories:
          type: array
          items:
            type: string
          example: [technology, adventure, fabric]
          description: A list of the featured categories that the project has.
        additional_categories:
          type: array
          items:
            type: string
          description: A list of additional categories that the project also has. These are supplementary to the featured categories, and does not include them again.
          example: [technology, adventure, fabric]
        environment:
          type: array
          items:
            $ref: '#/components/schemas/EnvironmentEnum'
          description: 'All the environments that versions of this project support. Not in any particular order, we recommend using the environment information on a version instead. For an explanation of each environment, see the blog post here: https://modrinth.com/news/article/new-environments/#new-system'
          example: [client_and_server]
        game_versions:
          type: array
          items:
            type: string
          example: ['1.19', '1.19.1', '1.19.2', '1.19.3']
          description: A list of all of the game versions supported by the project
        loaders:
          type: array
          items:
            type: string
          example: ['neoforge', 'fabric']
          description: A list of all of the loaders supported by the project. These vary based on project type.
        versions:
          type: array
          items:
            type: string
          example: [IIJJKKLL, QQRRSSTT]
          description: A list of the version IDs of the project
        license:
          $ref: '#/components/schemas/ProjectLicense'
        published:
          type: string
          format: ISO-8601
          description: The date the project was created
        updated:
          type: string
          format: ISO-8601
          description: The date the latest version of the project was created
        downloads:
          type: integer
          description: The total number of downloads of the project
        followers:
          type: integer
          description: The total number of users following the project
        gallery:
          type: array
          items:
            $ref: '#/components/schemas/GalleryImage'
          description: A list of images that have been uploaded to the project's gallery
        thread_id:
          type: string
          example: TTUUVVWW
          description: The ID of the moderation thread associated with this project
        monetization_status:
          type: string
          enum: [monetized, demonetized, force-demonetized]
        slug:
          type: string
          description: "The slug of a project, used for vanity URLs. Regex: ```^[\\w!@$()`.+,\"\\-']{3,64}$```"
          example: my_project
          nullable: true
        organization:
          type: string
          description: The ID of the organization that owns this project
          example: AABBCCDD
          nullable: true
        requested_status:
          type: string
          enum: [approved, archived, unlisted, private, draft]
          description: The requested status when submitting for review or scheduling the project for release. Approved status refers to "Public" visibility.
          nullable: true
        approved:
          type: string
          format: ISO-8601
          description: The date the project was first published
          nullable: true
        queued:
          type: string
          format: ISO-8601
          description: The date the project's status was submitted to moderators for review
          nullable: true
        icon_url:
          type: string
          example: https://cdn.modrinth.com/data/AABBCCDD/b46513nd83hb4792a9a0e1fn28fgi6090c1842639.png
          description: The URL of the project's icon
          nullable: true
        raw_icon_url:
          type: string
          example: https://cdn.modrinth.com/data/AABBCCDD/b46513nd83hb4792a9a0e1fn28fgi6090c1842639-raw.png
          description: The URL of the project's icon without CDN transforms applied
          nullable: true
        color:
          type: integer
          example: 8703084
          description: The RGB color of the project, automatically generated from the project icon
          nullable: true
        issues_url:
          type: string
          description: An optional link to where to submit bugs or issues with the project
          example: https://github.com/modrinth/code/issues
          nullable: true
        source_url:
          type: string
          description: An optional link to the source code of the project
          example: https://github.com/modrinth/code
          nullable: true
        wiki_url:
          type: string
          description: An optional link to the project's wiki page or other relevant information
          example: https://github.com/modrinth/code/wiki
          nullable: true
        discord_url:
          type: string
          description: An optional invite link to the project's discord.
          example: https://discord.gg/modrinth
          nullable: true
        donation_urls:
          type: array
          items:
            $ref: '#/components/schemas/ProjectDonationURL'
          description: A list of donation links for the project
          nullable: true
        client_side:
          type: string
          enum: [required, optional, unsupported, unknown]
          description: Deprecated - use `environment` instead.
          example: required
          deprecated: true
        server_side:
          type: string
          enum: [required, optional, unsupported, unknown]
          description: Deprecated - use `environment` instead.
          example: optional
          deprecated: true
        body_url:
          type: string
          deprecated: true
          default: null
          description: Deprecated - The link to the long description of the project. Always null, only kept for legacy compatibility.
          example: null
          nullable: true
        moderator_message:
          $ref: '#/components/schemas/ModeratorMessage'
      required:
        - id
        - team
        - title
        - description
        - body
        - status
        - project_type
        - categories
        - additional_categories
        - environment
        - game_versions
        - loaders
        - versions
        - license
        - published
        - updated
        - downloads
        - followers
        - gallery
        - thread_id
        - monetization_status
        - client_side
        - server_side
...
    InvalidInputError:
      type: object
      properties:
        error:
          type: string
          description: The name of the error
          example: 'invalid_input'
        description:
          type: string
          description: The contents of the error
          example: 'Error while parsing multipart payload'
      required:
        - error
        - description
    AuthError:
      type: object
      properties:
        error:
          type: string
          description: The name of the error
          example: 'unauthorized'
        description:
          type: string
          description: The contents of the error
          example: 'Authentication Error: Invalid Authentication Credentials'
      required:
        - error
        - description
    # Other
...
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
...
paths:
...
  /project/{id|slug}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdentifier'
    get:
      summary: Get a project
      operationId: getProject
      tags:
        - projects
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: The requested item(s) were not found or no authorization to access the requested item(s)
    patch:
      summary: Modify a project
      operationId: modifyProject
      tags:
        - projects
      security:
        - TokenAuth: ['PROJECT_WRITE']
      requestBody:
        description: 'Modified project fields'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditableProject'
      responses:
        '204':
          description: Expected response to a valid request
        '401':
          description: Incorrect token scopes or no authorization to access the requested item(s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: The requested item(s) were not found or no authorization to access the requested item(s)
    delete:
      summary: Delete a project
      operationId: deleteProject
      tags:
        - projects
      security:
        - TokenAuth: ['PROJECT_DELETE']
      responses:
        '204':
          description: Expected response to a valid request
        '400':
          description: Request was invalid, see given error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidInputError'
        '401':
          description: Incorrect token scopes or no authorization to access the requested item(s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
...
  /version/{id}:
    parameters:
      - $ref: '#/components/parameters/VersionIdentifier'
    get:
      summary: Get a version
      operationId: getVersion
      tags:
        - versions
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '404':
          description: The requested item(s) were not found or no authorization to access the requested item(s)
    patch:
      summary: Modify a version
      operationId: modifyVersion
      tags:
        - versions
      security:
        - TokenAuth: ['VERSION_WRITE']
      requestBody:
        description: 'Modified version fields'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditableVersion'
      responses:
        '204':
          description: Expected response to a valid request
        '401':
          description: Incorrect token scopes or no authorization to access the requested item(s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: The requested item(s) were not found or no authorization to access the requested item(s)
    delete:
      summary: Delete a version
      operationId: deleteVersion
      tags:
        - versions
      security:
        - TokenAuth: ['VERSION_DELETE']
      responses:
        '204':
          description: Expected response to a valid request
        '401':
          description: Incorrect token scopes or no authorization to access the requested item(s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: The requested item(s) were not found or no authorization to access the requested item(s)
  /project/{id|slug}/version/{id|number}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdentifier'
      - name: id|number
        in: path
        required: true
        description: The version ID or version number
        schema:
          type: string
          example: [IIJJKKLL]
    get:
      summary: Get a version given a version number or ID
      description: Please note that, if the version number provided matches multiple versions, only the **oldest matching version** will be returned.
      operationId: getVersionFromIdOrNumber
      tags:
        - versions
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '404':
          description: The requested item(s) were not found or no authorization to access the requested item(s)
...
  /versions:
    parameters:
      - in: query
        name: ids
        description: The IDs of the versions
        schema:
          type: string
          example: '["AABBCCDD", "EEFFGGHH"]'
        required: true
    get:
      summary: Get multiple versions
      operationId: getVersions
      tags:
        - versions
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Version'
...